requirements.txt
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
